frontend een beetje opleuken

+++ frontend/app/components/caretaker/remove.js
@@ -8,11 +8,13 @@ export default class RemoveComponent extends Component {
 
   @action
   async removeCaretaker(caretaker) {
+   if (window.confirm('Are you sure you want to delete ' + caretaker.name)){
     try {
       await caretaker.destroyRecord();
     } catch (error) {
       console.error('Error removing caretaker:', error);
     }
     this.router.transitionTo('/caretakers');
+   }
   }
 }
